requirements.txt
BayesAgg-MTL==1.0.0
Brotli==1.0.9
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==2.0.4
cryptography==41.0.7
filelock==3.13.1
fsspec==2024.3.1
gmpy2==2.1.2
huggingface-hub==0.22.2
idna==3.4
Jinja2==3.1.3
joblib==1.3.2
MarkupSafe==2.1.3
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
networkx==3.1
numpy==1.26.3
packaging==24.0
pandas==2.2.0
Pillow==10.0.1
pip==23.3.1
pycparser==2.21
pyOpenSSL==23.2.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.4
PyYAML==6.0.1
requests==2.31.0
safetensors==0.4.3
scikit-learn==1.4.0
scipy==1.12.0
setuptools==68.2.2
six==1.16.0
sympy==1.12
threadpoolctl==3.2.0
timm==0.9.16
torch==2.1.1
torchvision==0.16.1
tqdm==4.66.1
triton==2.1.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==1.26.18
wheel==0.41.2
